modified: npm scripts
